chore(.gitignore): update gitignore rules chore(install-monitoring.yml): update playbook to install monitoring agent chore(outputs.tf): update output descriptions and values
The `.gitignore` file has been updated to include the `ansible/inventory` file, which was previously not ignored. This ensures that the inventory file is not tracked by Git.

In the `install-monitoring.yml` playbook, the following changes have been made:
- The playbook name has been updated to "Install Monitoring Agent on Microk8s control plane nodes" to provide more clarity.
- A new task has been added to reboot the control plane nodes.
- The playbook name for worker nodes has been updated to "Install Monitoring Agent on Microk8s worker nodes".
- A new task has been added to reboot the control plane nodes.

In the `outputs.tf` file, the following changes have been made:
- The output description and value for `bastion_ssh_command`