chore(INFO.md): add ibmcloud_api_key input variable chore(README.md): update guide with pre-requisites and deployment steps chore(providers.tf): add ibmcloud_api_key to provider configuration feat(tfvars-example): add ibmcloud_api_key variable chore(variables.tf): add ibmcloud_api_key variable feat(vpc-microk8s.png): add diagram image for the MicroK8s on IBM Cloud VPC guide
The changes in the files are made to add support for the `ibmcloud_api_key` input variable. This variable is required to deploy the resources and is used to authenticate with the IBM Cloud API. The README.md file is updated to include the pre-requisites for local and Cloud Shell environments, as well as the steps to deploy the cluster. The providers.tf file is modified to include the `ibmcloud_api_key` in the provider configuration. The tfvars-example file is